feat: add color, new style
