+++ src/routes/Main.css
@@ -5,13 +5,24 @@
 }
 
 .main-background {
-  background: linear-gradient(to bottom, #F59859, #FFFFFF 100%);
   height: 100dvh;
   width: 100vw;
   position: fixed;
   top: 0;
   left: 0;
   z-index: -1;
+}
+
+.main-background.good {
+  background: linear-gradient(to bottom, #A9D18E 0%, #BEDCA9 24%, #FFFFFF 100%);
+}
+
+.main-background.caution {
+  background: linear-gradient(to bottom, #F59859, #FFFFFF 100%);
+}
+
+.main-background.danger {
+  background: linear-gradient(to bottom, #D2544A, #FFFFFF 100%);
 }
 
 .main-percentage-section {
@@ -32,23 +43,34 @@
 }
 
 .main-status-section {
-  max-width: 262px;
   margin: auto;
-  margin-bottom: 40px;
+  margin-bottom: 20px;
 }
 
 .main-status-circle {
-  width: 100%;
+  max-width: 262px;
   aspect-ratio: 1;
   border: 2px solid #FFFFFF;
   border-radius: 100%;
-  background: radial-gradient(#FFFFFF 0%, #F59859 80%);
+  margin: auto;
   position: relative;
   display: flex;
   flex-direction: column;
   justify-content: center;
   position: relative;
   align-items: center;
+}
+
+.main-status-circle.good {
+  background: radial-gradient(#FFFFFF 0%, #A9D18E 80%);
+}
+
+.main-status-circle.caution {
+  background: radial-gradient(#FFFFFF 0%, #F59859 80%);
+}
+
+.main-status-circle.danger {
+  background: radial-gradient(#FFFFFF 0%, #D2544A 80%);
 }
 
 .main-status {
@@ -72,29 +94,55 @@
 }
 
 .main-factor-wrapper {
+  margin: auto;
+  max-width: 700px;
   display: flex;
-  margin: 0 auto;
   justify-content: center;
   align-items: center;
-  position: absolute;
-  bottom: 30px;
-  left: 50%;
-  transform: translateX(-50%);
   width: 100%;
+  background: rgba(255, 255, 255, 0.5);
+  border: 2px solid white;
+  position: relative;
+  margin-top: 46px;
+  padding: 24px 0 13px 0;
+  border-radius: 10px;
 }
 
 .main-factor-title {
   font-size: 16px;
-  font-weight: 300;
+  font-weight: 400;
+  border: 2px solid white;
+  padding: 5px 35px;
+  color: white;
+  position: absolute;
+  top: 0;
+  margin: 0;
+  transform: translateY(-50%);
+  border-radius: 10px;
+}
+
+.main-factor-title.good {
+  background-color: #A9D18E;
+}
+
+.main-factor-title.caution {
+  background-color: #F59859;
+}
+
+.main-factor-title.danger {
+  background-color: #D2544A;
 }
 
 .main-factor-list {
-  padding-left: 30px;
+  padding: 0;
+  list-style: none;
+  margin: 0;
 }
 
 .main-factor-item {
-  font-size: 16px;
-  text-align: left;
+  font-size: 20px;
+  font-weight: 400;
+  margin-bottom: 3px;
 }
 
 .main-column-container {
