+++ src/routes/Main.jsx
@@ -4,12 +4,14 @@ import { useEffect, useState } from 'react';
 import { getHealthInfoList } from '../api/healthApi';
 import { getUserInfo } from '../api/authApi';
 import { getStroke } from '../api/strokeApi';
+import { calcStrokeToText, formatFactorList } from '../services/strokeService';
 
 function Main(props) {
   const [healthInfo, setHealthInfo] = useState({});
   const [userInfo, setUserInfo] = useState({});
   const [strokeInfo, setStrokeInfo] = useState({});
   const [factorList, setFactorList] = useState([]);
+  const [result, setResult] = useState('');
   
   useEffect(() => {
     getUserInfo()
@@ -43,26 +45,8 @@ function Main(props) {
 
   useEffect(() => {
     setFactorList(formatFactorList(strokeInfo));
+    setResult(calcStrokeToText(strokeInfo));
   }, [strokeInfo]);
-
-  const formatFactorList = (stroke) => {
-    let factorList = [];
-
-    if (stroke.isWeight) {
-      factorList.push("높은 체중");
-    }
-    if (stroke.isAge) {
-      factorList.push("높은 연령대");
-    }
-    if (stroke.isBloodPressure) {
-      factorList.push("높은 혈압");
-    }
-    if (stroke.isHeartDisease) {
-      factorList.push("심혈관계 질환 보유");
-    }
-    
-    return factorList;
-  }
 
   return (
     <main className="main">
@@ -70,23 +54,23 @@ function Main(props) {
         <p><span className="main-user-name">{userInfo.name}</span>님의 뇌졸중 발병 확률은</p>
         <p><span className="main-percentage">{(strokeInfo.probability * 100).toFixed(0)}</span>% 입니다.</p>
       </section>
-      <section className="main-status-section">
-        <div className="main-status-circle">
+      <section className="main-status-section ">
+        <div className={"main-status-circle " + (result === '양호' ? 'good' : result === '주의' ? 'caution' : 'danger')}>
           <h1 className="main-status">
-            주의
+            {result}
           </h1>
-          <h1 className="main-status-inner">주의</h1>
-          <div className="main-factor-wrapper">
-            <h3 className="main-factor-title">주요 요인</h3>
+          <h1 className="main-status-inner">{result}</h1>
+        </div>
+        <div className="main-factor-wrapper">
+            <h3 className={"main-factor-title " + (result === '양호' ? 'good' : result === '주의' ? 'caution' : 'danger')}>주요 요인</h3>
             <ul className="main-factor-list">
               {
-                factorList.map((item, i) => {
+                factorList.length > 0 ? factorList.map((item, i) => {
                   return <li className="main-factor-item" key={i}>{item}</li>
-                })
+                }) : <li className="main-factor-item">요인이 없습니다.</li>
               }
             </ul>
           </div>
-        </div>
       </section>
       <section className="main-column-container">
         <h3 className="main-column-container-title"><span className="main-user-name">{userInfo.name}</span>님에게 추천드리는 조언 칼럼</h3>
@@ -99,7 +83,7 @@ function Main(props) {
         <ColumnPost id={7} />
         <ColumnPost id={8} />
       </section>
-      <div className="main-background"></div>
+      <div className={"main-background " + (result === '양호' ? 'good' : result === '주의' ? 'caution' : 'danger')}></div>
     </main>
   )
 }
